pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["Ayomide <ayomidet905@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django-jazzmin = "^2.6.0"
cloudinary = "^1.36.0"
Pillow = "^10.1.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
django = "^5.2.1"
djangorestframework = "^3.14.0"
django-unfold = "^0.78.0"
django-ckeditor-5 = "^0.2.19"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
dj-database-url = "^2.1.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
